feat: add `MERGE_COMMIT_MESSAGE_CUT_BODY_REGEX` option
Can be used to prevent boilerplate parts of the PR message from making
itself into the commit history of the project (and potentially
subsequently polluting the CHANGELOG when using conventional commits).

Motivation: we have a project with a PR template that ends in a
licensing agreement. We squash-merge PRs to `master` and then use
`conventional-commits` to automatically generate a CHANGELOG from the
commits. I don't want the licensing footer to end up in the commit
message, since:

1) It has no business there; and
2) If the commit contains the term 'BREAKING CHANGE', everything
following that text will be copied into the changelog, including
the licensing footer.

This feature allows us to clean up our commit history.